SIMDify tokenization in streaming jpegli encoder.
Benchmark before (speed only, bo butteraugli):
``
Encoding                                  kPixels    Bytes          BPP  E MP/s  D MP/s
---------------------------------------------------------------------------------------
jpeg:enc-jpegli:p0:q90                      13270  3673582    2.2145939  84.327 194.893
jpeg:enc-jpegli:p0:fix:q90                  13270  3825338    2.3060790  91.094 189.485
jpeg:enc-jpegli:noaq:p0:fix:q90             13270  4280672    2.5805740  95.646 174.698
jpeg:enc-jpegli:noaq:p0:fix:q90:yuv420      13270  3595502    2.1675239 130.254 239.438
jpeg:enc-jpegli:noaq:p0:fix:q75:yuv420      13270  2165627    1.3055335 164.130 299.911
jpeg:q90                                    13270  4838710    2.9169834  66.587 170.812
jpeg:fix:q90                                13270  4945178    2.9811670 131.770 161.121
jpeg:fix:q90:yuv420                         13270  3736238    2.2523657 197.548 225.414
jpeg:fix:q75:yuv420                         13270  2200650    1.3266469 229.953 280.381
Aggregate:                                  13270  3561332    2.1469246 122.606 210.337
```

Benchmark after:
```
Encoding                                  kPixels    Bytes          BPP  E MP/s  D MP/s
---------------------------------------------------------------------------------------
jpeg:enc-jpegli:p0:q90                      13270  3673582    2.2145939  84.407 194.442
jpeg:enc-jpegli:p0:fix:q90                  13270  3825338    2.3060790 101.582 186.985
jpeg:enc-jpegli:noaq:p0:fix:q90             13270  4280672    2.5805740 113.451 173.329
jpeg:enc-jpegli:noaq:p0:fix:q90:yuv420      13270  3595502    2.1675239 168.617 225.264
jpeg:enc-jpegli:noaq:p0:fix:q75:yuv420      13270  2165627    1.3055335 187.170 284.319
jpeg:q90                                    13270  4838710    2.9169834  65.332 169.069
jpeg:fix:q90                                13270  4945178    2.9811670 130.767 161.496
jpeg:fix:q90:yuv420                         13270  3736238    2.2523657 198.441 229.697
jpeg:fix:q75:yuv420                         13270  2200650    1.3266469 233.308 285.668
Aggregate:                                  13270  3561332    2.1469246 131.972 207.821
```
